--- component-library/components/atoms/checkbox/checkbox.scss
@@ -1,98 +1,0 @@
-@import '../../utils/style.scss';
-
-
-:global(.layout-g) {
-
-    input[type="checkbox"], input[type="radio"] {
-        opacity: 0;
-        width: 20px;
-        + label {
-            position: relative;
-        }
-
-        /*Unhide the mark on the checked state*/
-        &:checked {
-            + label::after {
-                content: "";
-            }
-        }
-
-        /*Adding focus styles on the outer-box of the fake checkbox*/
-        &:focus {
-            + label::before {
-                outline: 1px dotted #777777;
-                outline-offset: 2px;
-            }
-        }
-        &:disabled {
-            +label, &::before, &::after   {
-                color: $disabled-gray;
-                border-color: $disabled-gray;
-                background-color: $disabled-gray;
-            }
-        }
-    }
-     /* Specific styles for checkboxes*/
-    input[type="checkbox"] {
-        + label {
-            &::before {
-                /*Outer-box*/
-                content: "";
-                display: inline-block;
-                height: 20px;
-                width: 20px;
-                @include border-top-radius($global-radius);
-                @include border-bottom-radius($global-radius);
-                border: 2px solid $black;
-                position: absolute;
-                top: -2px;
-                @include left(-25px);
-            }
-            &::after {
-                /*Checkmark*/
-                content: none;
-                display: inline-block;
-                height: 8px;
-                width: 8px;
-                border: 2px solid $black;
-                background-color: $black;
-                position: absolute;
-                @include left(-19px);
-                top: 4px;
-            }
-        }
-    }
-
-    /*Specific styles for radio-buttons*/
-    input[type="radio"] {
-        + label {
-            &::before {
-                /*Outer-box*/
-                content: "";
-                display: inline-block;
-                height: 20px;
-                width: 20px;
-                @include border-top-radius(50%);
-                @include border-bottom-radius(50%);
-                border: 2px solid $black;
-                position: absolute;
-                top: 0;
-                @include left(-25px);
-                background-color: $bg-white;
-            }
-            &::after {
-                /*Checkmark*/
-                content: none;
-                display: inline-block;
-                height: 6px;
-                width: 6px;
-                border: 2px solid;
-                background-color: $black;
-                border-radius: 50%;
-                position: absolute;
-                @include left(-18px);
-                top: 7px;
-            }
-        }
-    }
-}
