+++ component-library/components/atoms/formelements/checkbox/stories.js
@@ -3,7 +3,7 @@ import { storiesOf } from "@storybook/react";
 
 import Checkbox from "./checkbox";
 
-storiesOf("1 Atoms/Checkbox", module)
+storiesOf("1 Atoms/Form Elements/Checkbox", module)
     .add("checkbox", () => (
         <Checkbox checkType="CHECKBOX" id="a">
             Learn more
