+++ component-library/components/atoms/formelements/dropdown/dropdown.js
@@ -1,7 +1,7 @@
 import * as React from "react";
 import PropTypes from "prop-types";
 import classnames from "classnames";
-import dropdown from "./dropdown.scss";
+import dropdown from "../forms.scss";
 
 const propTypes = {
     className: PropTypes.string,
