+++ component-library/components/atoms/formelements/input/input.js
@@ -1,7 +1,7 @@
 import * as React from "react";
 import PropTypes from "prop-types";
 import classnames from "classnames";
-import input from "./input.scss";
+import input from "../forms.scss";
 
 const InputType = {
     TEXT: "text",
