--- component-library/components/atoms/label/label.scss
@@ -1,32 +1,0 @@
-@import '../../utils/style.scss';
-
-    :global(.layout-g) {
-        label,
-    .form-label {
-        @include typeGamma;
-
-        svg.iconic {
-            width: 0.8em;
-            height: 0.8em;
-            fill: $dark-gray;
-            @include margin-left($tiny-spacing);
-        }
-
-        &.error {
-            color: $error-red;
-        }
-
-        &.required {
-
-            &:after {
-                content: '\2217';
-                position: relative;
-                top: -2px;
-                color: $error-red;
-                font-weight: normal;
-                font-size: 1.05em;
-                @include margin-left(4px);
-            }
-        }
-    }
-}
